+++ db/migrate/20250914134235_add_employee_type_ref_to_employees.rb
@@ -1,0 +1,5 @@
+class AddEmployeeTypeRefToEmployees < ActiveRecord::Migration[8.0]
+  def change
+    add_column :employees, :employee_type_ref_id, :integer
+  end
+end