+++ db/schema.rb
@@ -10,7 +10,33 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.0].define(version: 2025_09_13_182804) do
+ActiveRecord::Schema[8.0].define(version: 2025_09_14_145333) do
+  create_table "employees", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
+    t.string "user"
+    t.string "references"
+    t.string "firstname"
+    t.string "middlename"
+    t.string "lastname"
+    t.text "address1"
+    t.text "address2"
+    t.string "city"
+    t.string "mobile_number"
+    t.date "birthdate"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "employee_type_ref_id"
+    t.bigint "user_id", null: false
+    t.index ["user_id"], name: "index_employees_on_user_id"
+  end
+
+  create_table "ref_codes", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
+    t.string "name"
+    t.string "code"
+    t.string "code_type"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
   create_table "users", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
@@ -22,4 +48,6 @@ ActiveRecord::Schema[8.0].define(version: 2025_09_13_182804) do
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
+
+  add_foreign_key "employees", "users"
 end