+++ test/models/ref_code_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class RefCodeTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end